Moved ITreeModel & ITreeTimer to TreeDiagram.Interfaces
