+++ Railgun/Core/Containers/ITimerContainer.cs
@@ -1,4 +1,5 @@
 ﻿using System.Threading.Tasks;
+using TreeDiagram.Interfaces;
 using TreeDiagram.Models;
 
 namespace Railgun.Core.Containers
