+++ TreeDiagram/Models/ConfigBase.cs
@@ -1,5 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using TreeDiagram.Interfaces;
 
 namespace TreeDiagram.Models
 {
