+++ TreeDiagram/Models/TimerBase.cs
@@ -1,6 +1,7 @@
 using System;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using TreeDiagram.Interfaces;
 
 namespace TreeDiagram.Models
 {
